requirements.txt
streamlit==1.37.1
pandas==2.2.2
plotly==5.22.0
paho-mqtt==1.6.1
streamlit-autorefresh